+++ neodos-kernel/src/drivers/mod.rs
@@ -4,6 +4,7 @@ pub mod boot_ahci;
 pub mod block;
 pub mod boot_loader;
 pub mod builtin_drivers;
+pub mod caps;
 pub mod dependency;
 pub mod driver_loader;
 pub mod driver_runtime;